Extract _logging.py from _legacy.py
Moves rotate_logs, initialize_logging, redact_string, and ts_log (along
with LOG_DIR, LOG_FILE, MAX_LOG_BACKUPS constants) into a dedicated
modules/helpers/_logging.py module.

_legacy.py now imports ts_log from _logging so its internal call sites
continue to work. All ~100 external call sites use helpers.ts_log(...)
which resolves via helpers/__init__.py re-exports.

Reduces _legacy.py from 1613 to 1511 lines.
