+++ modules/helpers/__init__.py
@@ -50,3 +50,4 @@ from ._install_mode import *  # noqa: F403
 from ._kometa_version import *  # noqa: F403
 from ._pid import *  # noqa: F403
 from ._paths import *  # noqa: F403
+from ._logging import *  # noqa: F403
